+++ server/src/test/java/org/opensearch/index/snapshots/blobstore/FileInfoTests.java
@@ -169,7 +169,7 @@ public class FileInfoTests extends OpenSearchTestCase {
         );
         int numBytes = 0;
         for (int i = 0; i < info.numberOfParts(); i++) {
-            numBytes += info.partBytes(i);
+            numBytes += (int) info.partBytes(i);
         }
         assertEquals(numBytes, 36);
 
@@ -180,7 +180,7 @@ public class FileInfoTests extends OpenSearchTestCase {
         );
         numBytes = 0;
         for (int i = 0; i < info.numberOfParts(); i++) {
-            numBytes += info.partBytes(i);
+            numBytes += (int) info.partBytes(i);
         }
         assertEquals(numBytes, 35);
         final int numIters = randomIntBetween(10, 100);
@@ -189,7 +189,7 @@ public class FileInfoTests extends OpenSearchTestCase {
             info = new BlobStoreIndexShardSnapshot.FileInfo("foo", metadata, new ByteSizeValue(randomIntBetween(1, 1000)));
             numBytes = 0;
             for (int i = 0; i < info.numberOfParts(); i++) {
-                numBytes += info.partBytes(i);
+                numBytes += (int) info.partBytes(i);
             }
             assertEquals(numBytes, metadata.length());
         }
